+++ index.js
@@ -6,7 +6,7 @@
  */
 
 import { AppRegistry } from 'react-native'
-import App from './src/app'
+import App from './src/App'
 
 AppRegistry.registerComponent('chronomintrn', () => {
   return App
